[eas-cli] merge everything from actions/new to actions delete unused actions
delete old inner class

remove uneeded dist cert utils

merge dist cert utils

remove unused functions of pprofile utils

pop out new directory
