+++ packages/eas-cli/src/credentials/ios/actions/ConfigureProvisioningProfile.ts
@@ -2,91 +2,109 @@ import assert from 'assert';
 import chalk from 'chalk';
 import ora from 'ora';
 
-import Log from '../../../log';
-import { Action, CredentialsManager } from '../../CredentialsManager';
-import { Context } from '../../context';
 import {
-  DistributionCertificate,
-  ProvisioningProfileStoreInfo,
-} from '../appstore/Credentials.types';
-import { AppLookupParams } from '../credentials';
-import { selectProvisioningProfileFromAppleAsync } from './ProvisioningProfileUtils';
+  AppleDistributionCertificateFragment,
+  AppleProvisioningProfileFragment,
+} from '../../../graphql/generated';
+import Log from '../../../log';
+import { Context } from '../../context';
+import { AppLookupParams } from '../api/GraphqlClient';
+import { AppleProvisioningProfileMutationResult } from '../api/graphql/mutations/AppleProvisioningProfileMutation';
+import { ProvisioningProfileStoreInfo } from '../appstore/Credentials.types';
+import { AuthCtx } from '../appstore/authenticate';
+import { MissingCredentialsNonInteractiveError } from '../errors';
 
-export class UseExistingProvisioningProfile implements Action {
-  constructor(private app: AppLookupParams) {}
+export class ConfigureProvisioningProfile {
+  constructor(
+    private app: AppLookupParams,
+    private distributionCertificate: AppleDistributionCertificateFragment,
+    private originalProvisioningProfile: AppleProvisioningProfileFragment
+  ) {}
 
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    const selected = await selectProvisioningProfileFromAppleAsync(ctx, this.app.bundleIdentifier);
-    if (selected) {
-      const distCert = await ctx.ios.getDistributionCertificateAsync(this.app);
-      assert(distCert, 'missing distribution certificate');
-
-      await manager.runActionAsync(new ConfigureProvisioningProfile(this.app));
-    }
-  }
-}
-
-export class ConfigureProvisioningProfile implements Action {
-  constructor(private app: AppLookupParams) {}
-
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    const profile = await ctx.ios.getProvisioningProfileAsync(this.app);
-    if (!profile) {
-      throw new Error('No provisioning profile to configure');
-    }
-
-    if (!profile.provisioningProfileId) {
-      Log.warn("The provisioning profile we have on file cannot be validated on Apple's servers.");
-      return;
-    }
-
-    if (ctx.appStore.authCtx) {
-      const distCert = await ctx.ios.getDistributionCertificateAsync(this.app);
-      if (!distCert) {
-        Log.warn('There is no distribution certificate for this app.');
-        return;
-      }
-      const profilesFromApple = await ctx.appStore.listProvisioningProfilesAsync(
-        this.app.bundleIdentifier
+  public async runAsync(ctx: Context): Promise<AppleProvisioningProfileMutationResult | null> {
+    if (ctx.nonInteractive) {
+      throw new MissingCredentialsNonInteractiveError(
+        'Configuring Provisioning Profiles is only supported in interactive mode.'
       );
-      const profilesWithMatchingId = profilesFromApple.filter(
-        appleInfo => appleInfo.provisioningProfileId === profile.provisioningProfileId
-      );
-      if (!profilesWithMatchingId || profilesWithMatchingId.length < 1) {
-        Log.warn('This profile is no longer valid on Apple Developer Portal.');
-      }
+    }
+    const { developerPortalIdentifier, provisioningProfile } = this.originalProvisioningProfile;
+    if (!developerPortalIdentifier && !provisioningProfile) {
+      Log.warn("The provisioning profile we have on file cannot be configured on Apple's servers.");
+      return null;
+    }
 
-      await this.configureAndUpdateAsync(ctx, this.app, distCert, profilesWithMatchingId[0]);
-    } else {
+    if (!ctx.appStore.authCtx) {
       Log.warn(
         "Without access to your Apple account we can't configure provisioning profiles for you."
       );
       Log.warn('Make sure to recreate the profile if you selected a new distribution certificate.');
+      return null;
     }
+
+    const profilesFromApple = await ctx.appStore.listProvisioningProfilesAsync(
+      this.app.bundleIdentifier
+    );
+    const [matchingProfile] = profilesFromApple.filter(appleInfo =>
+      developerPortalIdentifier
+        ? appleInfo.provisioningProfileId === developerPortalIdentifier
+        : appleInfo.provisioningProfile === provisioningProfile
+    );
+    if (!matchingProfile) {
+      Log.warn(
+        `Profile ${
+          developerPortalIdentifier ? `${developerPortalIdentifier} ` : ''
+        }not found on Apple Developer Portal.`
+      );
+      return null;
+    }
+
+    return await this.configureAndUpdateAsync(ctx, ctx.appStore.authCtx, this.app, matchingProfile);
   }
 
   private async configureAndUpdateAsync(
     ctx: Context,
+    authCtx: AuthCtx,
     app: AppLookupParams,
-    distCert: DistributionCertificate,
     profileFromApple: ProvisioningProfileStoreInfo
-  ) {
+  ): Promise<AppleProvisioningProfileMutationResult> {
+    const {
+      developerPortalIdentifier,
+      certificateP12,
+      certificatePassword,
+      serialNumber,
+    } = this.distributionCertificate;
+    assert(
+      certificateP12 && certificatePassword,
+      'Distribution Certificate P12 and Password is required'
+    );
     // configure profile on Apple's Server to use our distCert
     const updatedProfile = await ctx.appStore.useExistingProvisioningProfileAsync(
       app.bundleIdentifier,
       profileFromApple,
-      distCert
+      {
+        certId: developerPortalIdentifier ?? undefined,
+        certP12: certificateP12,
+        certPassword: certificatePassword,
+        distCertSerialNumber: serialNumber,
+        teamId: authCtx.appleId,
+      }
     );
 
     const bundleIdTag = `(${app.bundleIdentifier})`;
-    const slugTag = `@${app.accountName}/${app.projectName}`;
+    const slugTag = `@${app.account.name}/${app.projectName}`;
     const projectTag = `${chalk.bold(slugTag)} ${chalk.dim(bundleIdTag)}`;
 
     const spinner = ora(`Updating Expo profile for ${projectTag}`).start();
     try {
-      // Update profile on EAS servers
-      await ctx.ios.updateProvisioningProfileAsync(app, updatedProfile);
+      const configuredProvisioningProfile = await ctx.newIos.updateProvisioningProfileAsync(
+        this.originalProvisioningProfile.id,
+        {
+          appleProvisioningProfile: updatedProfile.provisioningProfile,
+          developerPortalIdentifier: updatedProfile.provisioningProfileId,
+        }
+      );
       spinner.succeed(`Updated Expo profile for ${projectTag}`);
+      return configuredProvisioningProfile;
     } catch (error) {
       spinner.fail();
       throw error;
