+++ packages/eas-cli/src/credentials/ios/actions/CreateDistributionCertificate.ts
@@ -1,41 +1,16 @@
-import assert from 'assert';
-
 import Log from '../../../log';
-import { Action, CredentialsManager } from '../../CredentialsManager';
+import { Account } from '../../../user/Account';
 import { Context } from '../../context';
-import { IosDistCredentials } from '../credentials';
-import { displayIosUserCredentials } from '../utils/printCredentials';
+import { AppleDistributionCertificateMutationResult } from '../api/graphql/mutations/AppleDistributionCertificateMutation';
 import { provideOrGenerateDistributionCertificateAsync } from './DistributionCertificateUtils';
 
-export class CreateDistributionCertificateStandaloneManager implements Action {
-  constructor(private accountName: string) {}
+export class CreateDistributionCertificate {
+  constructor(private account: Account) {}
 
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    const action = new CreateDistributionCertificate(this.accountName);
-    await manager.runActionAsync(action);
-
-    Log.newLine();
-    displayIosUserCredentials(action.distCredentials);
-    Log.newLine();
-  }
-}
-
-export class CreateDistributionCertificate implements Action {
-  private _distCredentials?: IosDistCredentials;
-
-  constructor(private accountName: string) {}
-
-  public get distCredentials(): IosDistCredentials {
-    assert(this._distCredentials, 'distCredentials can be accessed only after runAsync');
-    return this._distCredentials;
-  }
-
-  public async runAsync(manager: CredentialsManager, ctx: Context): Promise<void> {
-    const distCert = await provideOrGenerateDistributionCertificateAsync(ctx, this.accountName);
-    this._distCredentials = await ctx.ios.createDistributionCertificateAsync(
-      this.accountName,
-      distCert
-    );
+  public async runAsync(ctx: Context): Promise<AppleDistributionCertificateMutationResult> {
+    const distCert = await provideOrGenerateDistributionCertificateAsync(ctx, this.account.name);
+    const result = await ctx.newIos.createDistributionCertificateAsync(this.account, distCert);
     Log.succeed('Created distribution certificate');
+    return result;
   }
 }
