+++ packages/eas-cli/src/credentials/ios/actions/__tests__/RemoveProvisioningProfile-test.ts
@@ -1,4 +1,4 @@
-import { createCtxMock } from '../../../../__tests__/fixtures-context';
+import { createCtxMock } from '../../../__tests__/fixtures-context';
 import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
 import { RemoveProvisioningProfiles } from '../RemoveProvisioningProfile';
